requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
Deprecated==1.2.14
Django==4.1
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
idna==3.4
jwcrypto==1.5.0
oauthlib==3.2.2
psycopg2==2.9.7
pycparser==2.21
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.4
wrapt==1.15.0
